build(maven): restructure project to build with maven
