+++ README.md
@@ -9,3 +9,25 @@ COMP 3005 A
 | Arthur Atangana  | 101005197      |
 | Rebecca Elliott  | 101199034      |
 | Braeden Kloke    | 100895984      |
+
+# Table of Contents
+1. [Build](#build)
+2. [Test](#test)
+3. [Run](#run)
+
+
+<a id="build"></a>
+# Build
+```
+mvn package
+```
+
+
+<a id="build"></a>
+# Test
+Tests are run manually by running the CLI!
+
+
+<a id="build"></a>
+# Run
+In IntelliJ, navigate to `src/main/java/ClubManagementSystem.java` and run the `main` method.
