+++ src/main/java/views/UsernameRegisterPrompt.java
@@ -1,6 +1,6 @@
-package main.views;
+package views;
 
-import main.ClubManagementCLI;
+import ui.ClubManagementCLI;
 
 /**
  * Class representing the prompt to register a new user's username.
